requirements.txt
amqp==5.0.9
anytree==2.8.0
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==3.2.10
django-bulk-update-or-create==0.3.0
django-celery-beat==2.2.1
django-timezone-field==4.2.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.10.0
kombu==5.2.3
mysqlclient==2.1.0
numpy==1.21.5
packaging==21.3
pandas==1.3.5
prompt-toolkit==3.0.24
pyparsing==3.0.6
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wincertstore==0.2
wrapt==1.13.3
zipp==3.6.0
